Get it to compiling fully

+++ src/element.cc
@@ -4,27 +4,25 @@
 
 #include "window.hh"
 
-#include "element.ii"
-
-#include "window.ii"
-
 
 element::element(element const & n_element):
-  m_x1(n_element.m_x1), m_y1(n_element.m_y1), m_x2(n_element.m_x2), m_y2(n_element.m_y2), m_z(n_element.m_z), m_owner(n_element.m_owner)
+  m_x1(n_element.m_x1), m_y1(n_element.m_y1), m_x2(n_element.m_x2),
+  m_y2(n_element.m_y2), m_z(n_element.m_z), m_owner(n_element.m_owner)
 {
 }
 
 
-element::element(int n_x1, int n_y1, int n_x2, int n_y2, int n_z, window & n_owner, bool n_visible):
+element::element(int16_t n_x1, int16_t n_y1, int16_t n_x2, int16_t n_y2,
+    int16_t n_z, window & n_owner, bool n_visible):
   m_x1(n_x1), m_y1(n_y1), m_x2(n_x2), m_y2(n_y2), m_z(n_z), m_owner(n_owner), m_visible(n_visible)
 {
 }
 
 
-void element::set_frame_pos(int x1, int y1)
+void element::set_frame_pos(int16_t x1, int16_t y1)
 {
-  int old_x1 = m_x1;
-  int old_y1 = m_y1;
+  int16_t old_x1 = m_x1;
+  int16_t old_y1 = m_y1;
 
   m_x2 = x1 + frame_width();
   m_y2 = y1 + frame_height();
@@ -37,7 +35,7 @@ void element::set_frame_pos(int x1, int y1)
 }
 
 
-void element::set_frame_size(int width, int height)
+void element::set_frame_size(int16_t width, int16_t height)
 {
   int old_width = frame_width();
   int old_height = frame_height();
@@ -51,9 +49,9 @@ void element::set_frame_size(int width, int height)
 }
 
 
-void element::set_frame_depth(int z)
+void element::set_frame_depth(int16_t z)
 {
-  int old_z = m_z;
+  int16_t old_z = m_z;
 
   m_z = z;
   if(m_visible)
@@ -63,7 +61,8 @@ void element::set_frame_depth(int z)
 }
 
 
-void element::set_frame(int x1, int y1, int x2, int y2, int z)
+void element::set_frame(int16_t x1, int16_t y1, int16_t x2, int16_t y2,
+    int16_t z)
 {
   int old_x1 = m_x1;
   int old_y1 = m_y1;
@@ -88,7 +87,7 @@ void element::repaint()
 }
 
 
-void element::repaint(int x1, int y1, int x2, int y2)
+void element::repaint(int16_t x1, int16_t y1, int16_t x2, int16_t y2)
 {
   if(m_visible)
   {
